Updated aboutUs page
