+++ src/components/aboutUs/About.js
@@ -1,14 +1,15 @@
+// components/AboutSection.jsx
 "use client";
 
 import React from "react";
-import { Box, Container, Typography, Button } from "@mui/material";
 import Image from "next/image";
+import { Box, Container, Typography, Button } from "@mui/material";
 import CountUp from "react-countup";
-import aboutImage from "../../assets/Images/homepage/aboutus.png";
+import aboutImage from "../../assets/Images/aboutUs/about.png"; // Update this to your actual image path
 
 const AboutSection = () => {
   return (
-    <Box sx={{ py: { xs: 4, md: 6 }, bgcolor: "#fff" }}> {/* Reduced vertical padding */}
+    <Box sx={{ py: { xs: 4, md: 6 }, bgcolor: "#fff" }}>
       <Container maxWidth="lg">
         <Box
           sx={{
@@ -16,7 +17,7 @@ const AboutSection = () => {
             alignItems: "center",
             justifyContent: "space-between",
             flexDirection: { xs: "column", md: "row" },
-            gap: { xs: 4, md: 6 }, // Reduced gap
+            gap: { xs: 4, md: 6 },
           }}
         >
           {/* LEFT: Image */}
@@ -33,7 +34,7 @@ const AboutSection = () => {
                 position: "relative",
                 borderRadius: "16px",
                 overflow: "hidden",
-                width: { xs: "250px", sm: "280px", md: "300px" }, // Smaller image
+                width: { xs: "250px", sm: "280px", md: "300px" },
                 height: { xs: "320px", sm: "360px", md: "380px" },
               }}
             >
@@ -50,10 +51,7 @@ const AboutSection = () => {
 
           {/* RIGHT: Content */}
           <Box sx={{ flex: 1, textAlign: { xs: "center", md: "left" } }}>
-            <Typography
-              variant="body2"
-              sx={{ color: "#2563eb", fontWeight: "bold", mb: 1 }}
-            >
+            <Typography variant="body2" sx={{ color: "#2563eb", fontWeight: "bold", mb: 1 }}>
               About Us
             </Typography>
 
@@ -63,7 +61,7 @@ const AboutSection = () => {
                 fontWeight: "bold",
                 mb: 2,
                 lineHeight: 1.3,
-                fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" }, // Reduced heading
+                fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" },
               }}
             >
               Transforming visions into reality with{" "}
@@ -72,15 +70,14 @@ const AboutSection = () => {
               </Box>
             </Typography>
 
-            <Typography variant="body2" sx={{ mb: 1, color: "#555", fontSize: "0.9rem" }}>
-              Prabhu Enterprises is a multi-service company offering
-              construction, interiors, events, real estate, logistics, digital
-              marketing, and home appliances.
+            <Typography variant="body2" sx={{ mb: 1, color: "#555", fontSize: { xs: "0.85rem", sm: "0.9rem" } }}>
+              Prabhu Enterprises is a multi-service company offering construction, interiors, events,
+              real estate, logistics, digital marketing, and home appliances.
             </Typography>
 
-            <Typography variant="body2" sx={{ mb: 2, color: "#555", fontSize: "0.9rem" }}>
-              Our goal is simple — to provide seamless, affordable, and premium
-              services that transform ideas into lasting results across India.
+            <Typography variant="body2" sx={{ mb: 2, color: "#555", fontSize: { xs: "0.85rem", sm: "0.9rem" } }}>
+              Our goal is simple — to provide seamless, affordable, and premium services that
+              transform ideas into lasting results across India.
             </Typography>
 
             {/* Button */}
@@ -108,45 +105,30 @@ const AboutSection = () => {
                 display: "flex",
                 flexWrap: "wrap",
                 justifyContent: { xs: "center", md: "flex-start" },
-                gap: { xs: 3, md: 4 }, // Reduced gap
+                gap: { xs: 3, md: 4 },
               }}
             >
               <Box textAlign="center">
-                <Typography
-                  sx={{
-                    fontWeight: "bold",
-                    fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" },
-                  }}
-                >
+                <Typography sx={{ fontWeight: "bold", fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" } }}>
                   <CountUp end={500} duration={3} />+
                 </Typography>
-                <Typography variant="body2" sx={{ color: "#555", fontSize: "0.8rem" }}>
+                <Typography variant="body2" sx={{ color: "#555", fontSize: { xs: "0.75rem", sm: "0.8rem" } }}>
                   Satisfied Customers
                 </Typography>
               </Box>
               <Box textAlign="center">
-                <Typography
-                  sx={{
-                    fontWeight: "bold",
-                    fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" },
-                  }}
-                >
+                <Typography sx={{ fontWeight: "bold", fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" } }}>
                   <CountUp end={10} duration={3} />+
                 </Typography>
-                <Typography variant="body2" sx={{ color: "#555", fontSize: "0.8rem" }}>
+                <Typography variant="body2" sx={{ color: "#555", fontSize: { xs: "0.75rem", sm: "0.8rem" } }}>
                   Years Experience
                 </Typography>
               </Box>
               <Box textAlign="center">
-                <Typography
-                  sx={{
-                    fontWeight: "bold",
-                    fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" },
-                  }}
-                >
+                <Typography sx={{ fontWeight: "bold", fontSize: { xs: "1.5rem", sm: "1.7rem", md: "2rem" } }}>
                   <CountUp end={40} duration={3} />+
                 </Typography>
-                <Typography variant="body2" sx={{ color: "#555", fontSize: "0.8rem" }}>
+                <Typography variant="body2" sx={{ color: "#555", fontSize: { xs: "0.75rem", sm: "0.8rem" } }}>
                   Projects Completed
                 </Typography>
               </Box>
